RDM-4824 - extracted to separate class to persist in a new transaction. applied to create case as well.
