+++ src/test/java/uk/gov/hmcts/ccd/domain/service/createevent/DefaultCreateEventOperationTest.java
@@ -1,25 +1,5 @@
 package uk.gov.hmcts.ccd.domain.service.createevent;
 
-import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-
-import static org.apache.http.HttpStatus.SC_OK;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.mockito.BDDMockito.given;
-import static org.mockito.Mockito.any;
-import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyZeroInteractions;
-import static uk.gov.hmcts.ccd.domain.model.std.EventBuilder.anEvent;
-import static uk.gov.hmcts.ccd.domain.service.common.TestBuildersUtil.CaseDataContentBuilder.newCaseDataContent;
-
 import com.fasterxml.jackson.databind.JsonNode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -27,43 +7,31 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.mockito.Spy;
 import org.springframework.http.ResponseEntity;
-import uk.gov.hmcts.ccd.data.casedetails.CaseAuditEventRepository;
-import uk.gov.hmcts.ccd.data.casedetails.CaseDetailsRepository;
-import uk.gov.hmcts.ccd.data.definition.CaseDefinitionRepository;
-import uk.gov.hmcts.ccd.data.user.UserRepository;
-import uk.gov.hmcts.ccd.domain.model.aggregated.IdamUser;
 import uk.gov.hmcts.ccd.domain.model.callbacks.AfterSubmitCallbackResponse;
-import uk.gov.hmcts.ccd.domain.model.callbacks.SignificantItem;
-import uk.gov.hmcts.ccd.domain.model.callbacks.SignificantItemType;
-import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
-import uk.gov.hmcts.ccd.domain.model.definition.CaseEvent;
-import uk.gov.hmcts.ccd.domain.model.definition.CaseState;
-import uk.gov.hmcts.ccd.domain.model.definition.CaseType;
-import uk.gov.hmcts.ccd.domain.model.definition.Jurisdiction;
-import uk.gov.hmcts.ccd.domain.model.definition.Version;
+import uk.gov.hmcts.ccd.domain.model.definition.*;
 import uk.gov.hmcts.ccd.domain.model.std.CaseDataContent;
 import uk.gov.hmcts.ccd.domain.model.std.Event;
 import uk.gov.hmcts.ccd.domain.model.std.validator.EventValidator;
-import uk.gov.hmcts.ccd.domain.service.callbacks.EventTokenService;
-import uk.gov.hmcts.ccd.domain.service.common.CaseDataService;
-import uk.gov.hmcts.ccd.domain.service.common.CaseService;
-import uk.gov.hmcts.ccd.domain.service.common.CaseTypeService;
-import uk.gov.hmcts.ccd.domain.service.common.EventTriggerService;
-import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
-import uk.gov.hmcts.ccd.domain.service.common.UIDService;
-import uk.gov.hmcts.ccd.domain.service.stdapi.AboutToSubmitCallbackResponse;
 import uk.gov.hmcts.ccd.domain.service.stdapi.CallbackInvoker;
-import uk.gov.hmcts.ccd.domain.service.util.TransactionHelper;
-import uk.gov.hmcts.ccd.domain.service.validate.ValidateCaseFieldsOperation;
-import uk.gov.hmcts.ccd.domain.types.sanitiser.CaseSanitiser;
 import uk.gov.hmcts.ccd.endpoint.exceptions.CallbackException;
-import uk.gov.hmcts.ccd.infrastructure.user.UserAuthorisation;
+
+import java.time.LocalDateTime;
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.apache.http.HttpStatus.SC_OK;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.*;
+import static uk.gov.hmcts.ccd.domain.model.std.EventBuilder.anEvent;
+import static uk.gov.hmcts.ccd.domain.service.common.TestBuildersUtil.CaseDataContentBuilder.newCaseDataContent;
 
 class DefaultCreateEventOperationTest {
 
-    private static final String USER_ID = "123";
     private static final String JURISDICTION_ID = "SSCS";
     private static final String CASE_TYPE_ID = "Claim";
     private static final String CASE_REFERENCE = "1234123412341236";
@@ -77,39 +45,11 @@ class DefaultCreateEventOperationTest {
     private static final String CALLBACK_URL = "http://sscs.reform.hmcts.net/callback";
 
     @Mock
-    private UserRepository userRepository;
-    @Mock
-    private CaseDetailsRepository caseDetailsRepository;
-    @Mock
-    private CaseDefinitionRepository caseDefinitionRepository;
-    @Mock
-    private CaseAuditEventRepository caseAuditEventRepository;
-    @Mock
-    private EventTriggerService eventTriggerService;
-    @Mock
-    private EventTokenService eventTokenService;
-    @Mock
-    private CaseDataService caseDataService;
-    @Mock
-    private CaseTypeService caseTypeService;
-    @Mock
     private EventValidator eventValidator;
-    @Mock
-    private CaseSanitiser caseSanitiser;
     @Mock
     private CallbackInvoker callbackInvoker;
     @Mock
-    private UIDService uidService;
-    @Mock
-    private SecurityClassificationService securityClassificationService;
-    @Mock
-    private ValidateCaseFieldsOperation validateCaseFieldsOperation;
-    @Mock
-    private CaseService caseService;
-    @Spy
-    private TransactionHelper transactionHelper;
-    @Mock
-    private UserAuthorisation userAuthorisation;
+    private CreateCaseEventService createEventService;
 
     @InjectMocks
     private DefaultCreateEventOperation createEventOperation;
@@ -147,15 +87,10 @@ class DefaultCreateEventOperationTest {
         caseType.setId(CASE_TYPE_ID);
         caseType.setJurisdiction(jurisdiction);
         caseType.setVersion(version);
+
         eventTrigger = new CaseEvent();
         eventTrigger.setPostState(POST_STATE);
-        final SignificantItem significantItem = new SignificantItem();
-        significantItem.setUrl("http://www.yahoo.com");
-        significantItem.setDescription("description");
-        significantItem.setType(SignificantItemType.DOCUMENT.name());
-        final AboutToSubmitCallbackResponse aboutToSubmitCallbackResponse = new AboutToSubmitCallbackResponse();
-        aboutToSubmitCallbackResponse.setSignificantItem(significantItem);
-        aboutToSubmitCallbackResponse.setState(Optional.empty());
+
         caseDetails = new CaseDetails();
         caseDetails.setCaseTypeId(CASE_TYPE_ID);
         caseDetails.setState(PRE_STATE_ID);
@@ -163,52 +98,15 @@ class DefaultCreateEventOperationTest {
         caseDetailsBefore = mock(CaseDetails.class);
         postState = new CaseState();
         postState.setId(POST_STATE);
-        IdamUser user = new IdamUser();
-        user.setId("123");
 
-        doReturn(caseType).when(caseDefinitionRepository).getCaseType(CASE_TYPE_ID);
-        doReturn(true).when(caseTypeService).isJurisdictionValid(JURISDICTION_ID, caseType);
-        doReturn(eventTrigger).when(eventTriggerService).findCaseEvent(caseType, EVENT_ID);
-        doReturn(true).when(uidService).validateUID(CASE_REFERENCE);
-        doReturn(caseDetails).when(caseDetailsRepository).lockCase(Long.valueOf(CASE_REFERENCE));
-        doReturn(true).when(eventTriggerService).isPreStateValid(PRE_STATE_ID, eventTrigger);
-        doReturn(caseDetails).when(caseDetailsRepository).set(caseDetails);
-        doReturn(postState).when(caseTypeService).findState(caseType, POST_STATE);
-        doReturn(user).when(userRepository).getUser();
-        doReturn(caseDetailsBefore).when(caseService).clone(caseDetails);
-        given(callbackInvoker.invokeAboutToSubmitCallback(any(),
-                                                          any(),
-                                                          any(),
-                                                          any(),
-                                                          any())).willReturn(aboutToSubmitCallbackResponse);
-    }
+        CreateCaseEventResult caseEventResult =  CreateCaseEventResult.caseEventWith()
+            .caseDetailsBefore(caseDetailsBefore)
+            .savedCaseDetails(caseDetails)
+            .eventTrigger(eventTrigger)
+            .build();
 
-    @Test
-    @DisplayName("should create copy of case before mutating it")
-    void shouldCreateCopyOfCaseBeforeMutation() {
-        createCaseEvent();
+        given(createEventService.createCaseEvent(CASE_REFERENCE, caseDataContent)).willReturn(caseEventResult);
 
-        verify(caseService).clone(caseDetails);
-    }
-
-    @Test
-    @DisplayName("should not interact with before case details copy")
-    void shouldNotInteractWithBeforeCaseDetails() {
-        createCaseEvent();
-
-        verifyZeroInteractions(caseDetailsBefore);
-    }
-
-    @Test
-    @DisplayName("should invoke about to submit callback")
-    void shouldInvokeAboutToSubmitCallback() {
-        createCaseEvent();
-
-        verify(callbackInvoker).invokeAboutToSubmitCallback(eventTrigger,
-            caseDetailsBefore,
-            caseDetails,
-            caseType,
-            IGNORE_WARNING);
     }
 
     @Test
@@ -223,7 +121,7 @@ class DefaultCreateEventOperationTest {
                 caseDetailsBefore,
                 caseDetails);
 
-        final CaseDetails caseDetails = createCaseEvent();
+        final CaseDetails caseDetails = createEventOperation.createCaseEvent(CASE_REFERENCE, caseDataContent);
 
         assertAll(
             () -> verify(callbackInvoker).invokeSubmittedCallback(eventTrigger, caseDetailsBefore, this.caseDetails),
@@ -243,18 +141,13 @@ class DefaultCreateEventOperationTest {
                 caseDetailsBefore,
                 caseDetails);
 
-        final CaseDetails caseDetails = createCaseEvent();
+        final CaseDetails caseDetails = createEventOperation.createCaseEvent(CASE_REFERENCE, caseDataContent);
 
         assertAll(
             () -> assertNull(caseDetails.getAfterSubmitCallbackResponse()),
             () -> assertThat(caseDetails.getCallbackResponseStatusCode(), is(SC_OK)),
             () -> assertThat(caseDetails.getCallbackResponseStatus(), is("INCOMPLETE_CALLBACK"))
         );
-    }
-
-    private CaseDetails createCaseEvent() {
-        return createEventOperation.createCaseEvent(CASE_REFERENCE,
-            caseDataContent);
     }
 
     private Map<String, JsonNode> buildJsonNodeData() {
